Added CMake files to build the unrar lib
To automatically build the "unrar" library from the 3rdparty
directory try this from the libarchive root directory:

  $ cd ..
  $ mkdir libarchive-build
  $ cd libarchive-build
  $ cmake ../libarchive -DENABLE_UNRAR=ON
  $ make

The -DENABLE_UNRAR=ON swtich will unlock the "unrar" library.
The "archive" and "archive_static" library have a dependency
on the "unrar" project, if the latter was enabled as shown
before.

To force building the "unrar" library, without any other project
you can type:

  $ make project_unrar

At the moment all the dependecies are artifical, meaning that
no part of libarchive actively compiles with or links against
the "unrar" library.